requirements.txt
arrow==1.4.0
asgiref==3.11.1
blessed==1.30.0
charset-normalizer==3.4.4
diff-match-patch==20241021
Django==5.2.11
django-import-export==4.4.0
django-picklefield==3.4.0
django-q2==1.9.0
et_xmlfile==2.0.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
pillow==12.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==3.5.3
reportlab==4.4.10
qrcode==7.4.2
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.6.0
